pyproject.toml
[tool.poetry]
name = "ecommerce-graphql"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "ecommerce_graphql"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
passlib = "^1.7.4"
fastapi-jwt-auth = "^0.5.0"
python-multipart = "^0.0.6"
pillow = "^9.3.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.178.0"}
asyncpg = "^0.27.0"
jinja2 = "^3.1.2"
hypercorn = "^0.14.3"
pyrogram = "^2.0.102"
tgcrypto = "^1.2.5"
lxml = "^4.9.2"
instagrapi = "^1.17.1"
aiohttp = {extras = ["speedups"], version = "^3.8.3"}
sqlalchemy = "2.0.15"
alchemy-graph = "^0.1.2"
aio-pika = "^9.0.7"
tinydb = "^4.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
